+++ test.py
@@ -4,7 +4,7 @@ def test_pytest():
     p = pmf.pack.JavaPack("name", version="3")
 
     @p.function()
-    def test_function(f: pmf.func_handler.JavaFuncHandler):
+    def test_function(f: pmf.func_handler.JavaFunctionHandler):
         pass
     print(p.funcs)
 test_pytest()